* upstream-master: (277 commits)
  THRIFT-2290 anyfile.go:10: imported and not used: "math"
  THRIFT-2322: serializer testcase helper file regenerated with most recent Thrift compiler
  THRIFT-2322: wrong Go code generated for assignment of an optional set
  THRIFT-2271 JavaScript: Support for Multiplexing Services Patch: David Sautter + jslint fixes
  THRIFT-2195 Delphi event handlers - fix for multiplex protocol
  THRIFT-2288 Go impl of Thrift JSON protocol wrongly writes/expects true/false for bools
  THRIFT-2232 IsSet* broken in Go
  THRIFT-2194 Node: Not actually prepending residual data in TFramedTransport.receiver
  THRIFT-2263 java: Always generate good hashCode for Java Patch: Andrew Gaul
  THRIFT-2274 java: TNonblockingServer and TThreadedSelectorServer do not close their channel selectors on exit and leak file descriptors Patch:  Thomas Kielbus
  THRIFT-2282 Incorect code generated for some typedefs
  THRIFT-2281 Generated send/recv code ignores errors returned by the underlying protocol
  THRIFT-2280 TJSONProtocol.Flush() does not really flush the transport
  THRIFT-2040 "uninitialized variable" warnings on MSVC/windows Patch: Konrad Grochowski
  THRIFT-2278 Buffered transport doesn't support writes > buffer size
  THRIFT-2279 TSerializer only returns the first 1024 bytes serialized
  THRIFT-2275 Fix memory leak in golang compact_protocol.
  THRIFT-2249:Add SMP Build option to thrift.spec (and three config defines) Client: build Patch: Martin Vogt
  THRIFT-2267:Should be able to choose socket family in Python TSocket Client: py Patch: Abraham Elmahrek
  THRIFT-2267:Should be able to choose socket family in Python TSocket Client: py Patch: Abraham Elmahrek
  ...

+++ lib/cpp/src/thrift/async/TEvhttpServer.cpp
@@ -17,8 +17,8 @@
  * under the License.
  */
 
-#include "TEvhttpServer.h"
-#include "TAsyncBufferProcessor.h"
+#include <thrift/async/TEvhttpServer.h>
+#include <thrift/async/TAsyncBufferProcessor.h>
 #include <thrift/transport/TBufferTransports.h>
 #include <evhttp.h>
 
@@ -99,7 +99,7 @@ int TEvhttpServer::serve() {
 
 
 TEvhttpServer::RequestContext::RequestContext(struct evhttp_request* req) : req(req)
-  , ibuf(new TMemoryBuffer(EVBUFFER_DATA(req->input_buffer), EVBUFFER_LENGTH(req->input_buffer)))
+  , ibuf(new TMemoryBuffer(EVBUFFER_DATA(req->input_buffer), static_cast<uint32_t>(EVBUFFER_LENGTH(req->input_buffer))))
   , obuf(new TMemoryBuffer())
 {}
